feat: add GitHub Actions workflow for automated testing
- Add test.yml workflow to run tests on push/PR
- Set up Java 17 and Python 3.11 environment
- Run pytest for all tests in tests/ directory
- Include code quality checks with flake8 and mypy
- Consolidate test files from root to tests/ directory
- Add shift pinning tests to test_models.py
