+++ .github/workflows/test.yml
@@ -1,0 +1,43 @@
+name: Run Tests
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    
+    steps:
+    - uses: actions/checkout@v4
+    
+    - name: Set up Java 17
+      uses: actions/setup-java@v4
+      with:
+        distribution: 'temurin'
+        java-version: '17'
+    
+    - name: Install uv
+      uses: astral-sh/setup-uv@v5
+      with:
+        enable-cache: true
+        cache-dependency-glob: "uv.lock"
+    
+    - name: Set up Python
+      run: |
+        uv python install 3.11
+    
+    - name: Install dependencies
+      run: |
+        uv sync --no-install-project
+    
+    - name: Run tests
+      run: |
+        uv run pytest tests/ -v
+    
+    - name: Run linters
+      run: |
+        uv run flake8 . || true
+        uv run mypy . || true
