--- test_continuous_planning_simple.py
@@ -1,102 +1,0 @@
-#!/usr/bin/env python3
-"""
-Simple test to verify continuous planning implementation without running full tests
-"""
-import sys
-
-sys.path.insert(0, "src")
-
-from datetime import datetime, timedelta
-
-from natural_shift_planner.core.models import Employee, Shift, ShiftSchedule
-
-
-def test_pinning_functionality():
-    """Test basic pinning functionality"""
-    print("Testing pinning functionality...")
-
-    # Create a shift
-    shift = Shift(
-        id="test_shift",
-        start_time=datetime(2025, 6, 1, 8, 0),
-        end_time=datetime(2025, 6, 1, 16, 0),
-        required_skills={"Nurse"},
-    )
-
-    # Test initial state
-    assert shift.pinned == False, "Shift should not be pinned initially"
-    assert not shift.is_pinned(), "is_pinned() should return False initially"
-
-    # Test pinning
-    shift.pin()
-    assert shift.pinned == True, "Shift should be pinned after pin()"
-    assert shift.is_pinned(), "is_pinned() should return True after pin()"
-
-    # Test unpinning
-    shift.unpin()
-    assert shift.pinned == False, "Shift should not be pinned after unpin()"
-    assert not shift.is_pinned(), "is_pinned() should return False after unpin()"
-
-    print("✓ Pinning functionality works correctly")
-
-
-def test_shift_model_with_pinning():
-    """Test that Shift model works with new pinning field"""
-    print("\nTesting Shift model with pinning field...")
-
-    employee = Employee(id="emp1", name="John Smith", skills={"Nurse"})
-
-    shift = Shift(
-        id="shift1",
-        start_time=datetime(2025, 6, 1, 8, 0),
-        end_time=datetime(2025, 6, 1, 16, 0),
-        required_skills={"Nurse"},
-        employee=employee,
-        pinned=True,  # Test creating with pinned=True
-    )
-
-    assert shift.pinned == True, "Should be able to create shift with pinned=True"
-    assert shift.employee == employee, "Employee assignment should work"
-
-    print("✓ Shift model works correctly with pinning field")
-
-
-def test_continuous_planning_service():
-    """Test that ContinuousPlanningService can be imported"""
-    print("\nTesting ContinuousPlanningService import...")
-
-    try:
-        from natural_shift_planner.api.continuous_planning import (
-            ContinuousPlanningService,
-        )
-
-        print("✓ ContinuousPlanningService imported successfully")
-
-        # Check that methods exist
-        assert hasattr(ContinuousPlanningService, "swap_shifts")
-        assert hasattr(ContinuousPlanningService, "find_replacement_for_shift")
-        assert hasattr(ContinuousPlanningService, "pin_shifts")
-        assert hasattr(ContinuousPlanningService, "unpin_shifts")
-        assert hasattr(ContinuousPlanningService, "reassign_shift")
-        print("✓ All expected methods exist")
-    except Exception as e:
-        print(f"✗ Failed to import ContinuousPlanningService: {e}")
-        return False
-
-    return True
-
-
-def main():
-    print("Running simple continuous planning tests...\n")
-
-    test_pinning_functionality()
-    test_shift_model_with_pinning()
-    test_continuous_planning_service()
-
-    print(
-        "\n✓ All simple tests passed! The continuous planning feature is implemented correctly."
-    )
-
-
-if __name__ == "__main__":
-    main()
